requirements.txt
pdfplumber==0.11.4
pypdf==4.3.1
pymupdf==1.24.10
groq==0.11.0
# groq 0.11.0 is not compatible with httpx 0.28+ (removed `proxies` arg)
httpx>=0.27.2,<0.28
python-dotenv==1.0.1
Pillow==10.4.0
rich==13.9.4
tqdm==4.66.5
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
